Moves tile to object
